requirements.txt
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-ollama
mysql-connector-python
python-dotenv
faiss-cpu
requests
mcp
playwright
beautifulsoup4
lxml

fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-multipart
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pymupdf
oss2
chromadb
